rename cli_flags
