+++ src/cli/remove.rs
